Add proof of concept for a reservation system on HasContainer
The current implementation uses container.total_requested to help
Activity to make sure there is actual content/space available when a
ship is making a run. This system is flawed for containers which are
used as both a destination and an origin for different activities.
Add a unit test demonstrating how a "RelayContainer" could work with a
reservation system instead of the current total_requested.