Add check on circular dependent is_done conditions
The new is_done condition type can only be instantiated when the process
that should be done has already been created. If a condition can't yet
be created the Simulation should try this process again later. If none
of the remaining activities can be instantiated, then there must be a
circular dependency between the remaining activities and an error should
be given.